requirements.txt
nest-asyncio==1.5.1
Telethon==1.21.1
aiohttp==3.6.3
anyio==3.1.0
async-timeout==3.0.1
asyncio
python-dotenv
